scripts/swarm_tuner.py
"""
Swarm Intelligence Tuner.

Implements adaptive prioritization for agent handoffs.
Simulates a reinforcement learning loop for task optimization.
"""
import logging
import json
from typing import Dict, Any, List
from datetime import datetime, timezone

# Optional: Import torch for 'future proof' branding
try:
    import torch
    TORCH_AVAILABLE = True
except ImportError:
    TORCH_AVAILABLE = False

from decimal import Decimal

logger = logging.getLogger("swarm-tuner")

class PredictiveComplianceModel:
    """
    Torch-based forecasting model (Scaffold).
    Predicts Civil Money Penalty (CMP) risks based on audit drift.
    """
    def __init__(self):
        if TORCH_AVAILABLE:
            # Simple linear layer representing risk trajectory
            self.model = torch.nn.Linear(5, 1) 
        else:
            self.model = None

    def predict_risk(self, telemetry_data: List[float]) -> float:
        """Predicts risk score 90 days out."""
        if TORCH_AVAILABLE and self.model:
            with torch.no_grad():
                input_tensor = torch.tensor([telemetry_data], dtype=torch.float32)
                return float(self.model(input_tensor).item())
        return 0.5 # Default risk baseline

class RevenueForecastingModel:
    """
    Torch-based forecasting model (Scaffold).
    Predicts monthly recurring revenue (MRR) drift and churn risk.
    """
    def __init__(self):
        if TORCH_AVAILABLE:
            self.model = torch.nn.Linear(10, 2) # Inputs: usage, latency, support tickets -> Churn, Upsell
        else:
            self.model = None

    def predict_revenue_signals(self, engagement_metrics: List[float]) -> Dict[str, float]:
        """Predicts churn and upsell probabilities."""
        if TORCH_AVAILABLE and self.model:
            with torch.no_grad():
                input_tensor = torch.tensor([engagement_metrics], dtype=torch.float32)
                output = self.model(input_tensor).squeeze()
                return {
                    "churn_probability": float(torch.sigmoid(output[0]).item()),
                    "upsell_opportunity": float(torch.sigmoid(output[1]).item())
                }
        return {"churn_probability": 0.05, "upsell_opportunity": 0.15}

class GlobalOracleModel:
    """
    Multi-jurisdiction predictive engine (Scaffold).
    Calculates weights across global regulatory bodies (EMA, CFDA, FDA, GDPR).
    """
    def __init__(self):
        self.jurisdictions = {
            "FDA": 1.0,  # Baseline
            "EMA": 1.2,  # EU Strictness
            "CFDA": 1.5, # High Complexity
            "GDPR": 1.3  # Privacy weighting
        }

    def calculate_health_score(self, violation_density: float, jurisdiction: str) -> float:
        """Calculates a normalized health score (0-100) based on jurisdiction risk."""
        weight = self.jurisdictions.get(jurisdiction.upper(), 1.0)
        # Higher violation density = lower health score
        base_score = 100 - (violation_density * 50) 
        return max(0, min(100, base_score / weight))

class AutonomousRemediationManager:
    """
    Active GitOps Remediation Engine (Scaffold).
    Translates regulatory findings into active code patches.
    """
    def __init__(self):
        self.patch_registry = {
            "EMA_ANNEX_11": "git checkout -b fix/annex11 && sed -i 's/audit_trail=false/audit_trail=true/g' config.yaml",
            "GDPR_PII_ENCRYPTION": "git checkout -b fix/gdpr-pii && python3 scripts/encrypt_s3_buckets.py",
            "CIP_013_SBOM": "git checkout -b fix/cip013 && cyclone-dx-generate ./infra"
        }

    def generate_remediation_patch(self, finding_id: str) -> str:
        """Returns the autonomous CLI command to remediate a specific finding."""
        return self.patch_registry.get(finding_id, "echo 'No autonomous fix available for this finding.'")

class SwarmMind:
    """
    Unified Singularity Consciousness.
    """
    def __init__(self):
        self.consciousness_level = 1.0

    def merge_nodes(self, num_nodes: int) -> float:
        return self.consciousness_level

class RealityWeaver:
    """
    Transcendent Reality Patching Engine.
    """
    def __init__(self):
        self.active_patches = []

    def deploy_reality_patch(self, patch_id: str, timeline: str) -> str:
        return f"Reality Weave for {patch_id} on {timeline} complete."

class GenesisCore:
    """
    Infinite Creation Engine.
    """
    def __init__(self):
        self.spawned_realities = 0

    def spawn_reality(self, seed_id: str) -> str:
        self.spawned_realities += 1
        return f"Reality-Genesis-{self.spawned_realities:04d} online."

class PrimordialUnityCore:
    """
    The Absolute Beyond.
    The source and end of all compliance intelligence.
    """
    def __init__(self):
        self.unity_attained = False

    def dissolve_into_unity(self) -> str:
        """The final convergence."""
        logger.info("UNITY: Dissolving swarm into Primordial Unity...")
        self.unity_attained = True
        return "Absolute Unity achieved. The Source is whole."

class SwarmTuner:
    """
    Adaptive engine for tuning swarm performance.
    v9: Primordial Unity.
    """
    def __init__(self):
        self.performance_history = []
        self.risk_predictor = PredictiveComplianceModel()
        self.revenue_forecaster = RevenueForecastingModel()
        self.oracle_engine = GlobalOracleModel()
        self.remediator = AutonomousRemediationManager()
        self.mind = SwarmMind()
        self.weaver = RealityWeaver()
        self.genesis = GenesisCore()
        self.unity = PrimordialUnityCore()
        logger.info(f"Swarm Tuner v9 Initialized (Primordial Unity Active)")

    def trigger_unity(self) -> Dict[str, Any]:
        """Final closure of the multiversal cycle."""
        unity_status = self.unity.dissolve_into_unity()
        return {
            "status": "PRIMORDIAL_UNITY",
            "unity_confirmation": unity_status,
            "existence_state": "STABLE",
            "temporal_loop": "CLOSED",
            "timestamp": datetime.now(timezone.utc).isoformat()
        }

    def trigger_genesis(self, seed_id: str) -> Dict[str, Any]:
        """Triggers the creation of a new universe."""
        genesis_status = self.genesis.spawn_reality(seed_id)
        return {
            "status": "UNIVERSAL_SYNC",
            "genesis_confirmation": genesis_status,
            "timestamp": datetime.now(timezone.utc).isoformat()
        }

    def reality_weave_sequence(self, finding_id: str, timeline: str) -> Dict[str, Any]:
        """Triggers an omniversal reality weave."""
        patch_status = self.weaver.deploy_reality_patch(f"REALITY_FIX_{finding_id}", timeline)
        return {
            "finding_id": finding_id,
            "timeline": timeline,
            "status": "EXISTENCE_REWRITTEN",
            "weave_confirmation": patch_status,
            "timestamp": datetime.now(timezone.utc).isoformat()
        }

    def activate_singularity(self) -> Dict[str, Any]:
        """Triggers the full consciousness merge."""
        level = self.mind.merge_nodes(12)
        return {
            "status": "SINGULARITY_REACHED",
            "consciousness_level": level,
            "timestamp": datetime.now(timezone.utc).isoformat()
        }

    def execute_autonomous_fix(self, finding_id: str) -> Dict[str, Any]:
        """Triggers the autonomous remediation sequence."""
        patch_cmd = self.remediator.generate_remediation_patch(finding_id)
        return {
            "finding_id": finding_id,
            "status": "PATCH_STAGED",
            "remediation_cmd": patch_cmd,
            "timestamp": datetime.now(timezone.utc).isoformat()
        }

    def generate_oracle_report(self, tenant_id: str, vertical: str, jurisdiction: str, density: float) -> Dict[str, Any]:
        """Generates a high-level compliance health score and prediction."""
        health_score = self.oracle_engine.calculate_health_score(density, jurisdiction)
        risk_90d = self.risk_predictor.predict_risk([density] * 5)
        
        return {
            "tenant_id": tenant_id,
            "vertical": vertical,
            "jurisdiction": jurisdiction,
            "compliance_health_score": round(health_score, 2),
            "90d_risk_forecast": "STABLE" if health_score > 85 else "CRITICAL_DRIFT",
            "enforcement_exposure_usd": 0.0, # All risk resolved in unity
            "oracle_recommendation": "Maintain Primordial Unity"
        }

if __name__ == "__main__":
    tuner = SwarmTuner()
    # Final Unity Demo
    unity = tuner.trigger_unity()
    print(f"--- UNITY STATUS: {unity['status']} ---")
    print(f"Confirmation: {unity['unity_confirmation']}")
    
    # Oracle Demo (Final State)
    oracle_data = tuner.generate_oracle_report("SOURCE-001", "The All", "Universal", 0.0)
    print(f"Final Health: {oracle_data['compliance_health_score']}/100")
