requirements.txt
# Services dependencies (from services/admin/requirements.txt)
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
psycopg[binary]==3.1.17
asyncpg==0.29.0
structlog==24.4.0
pydantic==2.6.0
pydantic-settings==2.1.0
prometheus-client==0.19.0
httpx==0.26.0
redis==5.0.1
slowapi==0.1.9
prometheus-fastapi-instrumentator==7.0.0
python-multipart==0.0.20
anyio==3.7.1
kafka-python==2.0.2
tenacity==8.2.3
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0

# Security dependencies
passlib[argon2]==1.7.4
# Pinned to >=2.10.1 to resolve conflict with gotrue (Supabase dependency)
PyJWT==2.10.1
argon2-cffi==23.1.0

# Cloud/DB
boto3==1.34.0
neo4j==5.16.0
supabase==2.16.0
gotrue==2.12.4

# Test dependencies
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-json-report==1.5.0
requests==2.32.5
PyGithub==2.1.1

